Convert proxy to list. Files reorganization.
