+++ scripts/proxy_scrapper/proxy_scrapper.py
@@ -16,18 +16,3 @@ def scrap_proxy():
                         sep=':',
                         header=False,
                         mode='a')
-
-
-def delete_duplicates():
-    try:
-        proxy_list = pd.read_csv('proxy_list.csv', sep=':')
-        proxy_list = proxy_list.drop_duplicates()
-        proxy_list.to_csv('proxy_list.csv',
-                          encoding='utf-8',
-                          index=False,
-                          sep=':',
-                          header=False,
-                          mode='w')
-    except FileNotFoundError:
-        scrap_proxy()
-        delete_duplicates()
